Move hash library
