+++ cmd/state-svc/internal/hash/file_hasher.go
