+++ cmd/state-svc/internal/hash/file_hasher_test.go
